+++ packages/builder/src/settings/pages/ai/UpsertAIConfigButton.svelte
@@ -6,6 +6,10 @@
   import { ActionButton, Modal } from "@budibase/bbui"
   import CustomConfigModal from "./CustomConfigModal.svelte"
   import BBAIConfigModal from "./BBAIConfigModal.svelte"
+  import { onMount } from "svelte"
+  import { admin, licensing } from "@/stores/portal"
+  import { API } from "@/api"
+  import PortalModal from "./PortalModal.svelte"
 
   export let row: AIConfigResponse
 
@@ -13,8 +17,24 @@
 
   let configModal: { show: () => void; hide: () => void }
   let openModal: boolean
+  let hasLicenseKey: boolean
 
   $: configModal?.show()
+
+  onMount(async () => {
+    try {
+      const license = $licensing.license
+      const isOfflineLicense = () => license && "identifier" in license
+      if (isOfflineLicense()) {
+        hasLicenseKey = true
+      } else {
+        const licenseKeyResponse = await API.getLicenseKey()
+        hasLicenseKey = !!licenseKeyResponse?.licenseKey
+      }
+    } catch {
+      //
+    }
+  })
 </script>
 
 <ActionButton
@@ -36,6 +56,16 @@
         config={row}
         provider={row.provider}
         type={row.configType}
+      />
+    {:else if !hasLicenseKey}
+      <PortalModal
+        confirmHandler={() => {
+          window.open($admin.accountPortalUrl, "_blank")
+          openModal = false
+        }}
+        cancelHandler={() => {
+          openModal = false
+        }}
       />
     {:else}
       <BBAIConfigModal config={row} type={row.configType} />
