+++ packages/builder/src/settings/pages/ai/configs.svelte
@@ -1,23 +1,13 @@
 <script lang="ts">
-  import { admin, aiConfigsStore, licensing } from "@/stores/portal"
+  import { aiConfigsStore } from "@/stores/portal"
   import { Button, Layout, Modal, notifications, Table } from "@budibase/bbui"
-  import type { AIConfigResponse } from "@budibase/types"
   import { AIConfigType, BUDIBASE_AI_PROVIDER_ID } from "@budibase/types"
   import { onMount } from "svelte"
-
   import CustomConfigModal from "./CustomConfigModal.svelte"
-  import BBAIConfigModal from "./BBAIConfigModal.svelte"
-  import PortalModal from "./PortalModal.svelte"
-  import { API } from "@/api"
   import AILogo from "./AILogo.svelte"
   import UpsertAIConfigButton from "./UpsertAIConfigButton.svelte"
 
   let configModal: { show: () => void; hide: () => void }
-  let portalModal: { show: () => void; hide: () => void }
-  let selectedModalConfig: AIConfigResponse | undefined
-  let selectedProvider: string | undefined
-
-  let hasLicenseKey: boolean
 
   $: completionConfigs = ($aiConfigsStore.customConfigs || []).filter(
     config => config.configType === AIConfigType.COMPLETIONS
@@ -86,33 +76,13 @@
     edit: { width: "100px", align: "Right" },
   }
 
-  function createAIConfig(provider?: string) {
-    if (
-      provider === BUDIBASE_AI_PROVIDER_ID &&
-      !$admin.cloud &&
-      !hasLicenseKey
-    ) {
-      portalModal.show()
-      return
-    }
-
-    selectedModalConfig = undefined
-    selectedProvider = provider
+  function createAIConfig() {
     configModal?.show()
   }
 
   onMount(async () => {
     try {
       await aiConfigsStore.fetch()
-
-      const license = $licensing.license
-      const isOfflineLicense = () => license && "identifier" in license
-      if (isOfflineLicense()) {
-        hasLicenseKey = true
-      } else {
-        const licenseKeyResponse = await API.getLicenseKey()
-        hasLicenseKey = !!licenseKeyResponse?.licenseKey
-      }
     } catch {
       notifications.error("Error fetching AI settings")
     }
@@ -160,34 +130,7 @@
 </Layout>
 
 <Modal bind:this={configModal}>
-  {#if selectedProvider !== BUDIBASE_AI_PROVIDER_ID}
-    <CustomConfigModal
-      config={selectedModalConfig}
-      provider={selectedProvider}
-      type={AIConfigType.COMPLETIONS}
-      on:hide={() => {
-        configModal.hide()
-      }}
-    />
-  {:else}
-    <BBAIConfigModal
-      config={selectedModalConfig}
-      type={AIConfigType.COMPLETIONS}
-      on:hide={() => {
-        configModal.hide()
-      }}
-    />
-  {/if}
-</Modal>
-
-<Modal bind:this={portalModal}>
-  <PortalModal
-    confirmHandler={() => {
-      window.open($admin.accountPortalUrl, "_blank")
-      portalModal.hide()
-    }}
-    cancelHandler={() => portalModal.hide()}
-  />
+  <CustomConfigModal type={AIConfigType.COMPLETIONS} />
 </Modal>
 
 <style>
